refactor common date utils
