+++ apps/admin-ui/src/component/notifications/page.tsx
@@ -25,7 +25,7 @@ import {
   type MutationUpdateBannerNotificationArgs,
   type MutationCreateBannerNotificationArgs,
   type MutationDeleteBannerNotificationArgs,
-  ErrorType,
+  type ErrorType,
 } from "common/types/gql-types";
 import { BANNER_NOTIFICATIONS_ADMIN_LIST } from "common/src/components/BannerNotificationsQuery";
 import { H1 } from "common/src/common/typography";
@@ -36,13 +36,13 @@ import { publicUrl } from "app/common/const";
 import Loader from "app/component/Loader";
 import { useNotification } from "app/context/NotificationContext";
 import { checkDate, checkTimeStringFormat } from "app/schemas";
-import ControlledDateInput from "../my-units/components/ControlledDateInput";
 import {
   valueForDateInput,
   valueForTimeInput,
   dateTime,
   parseDateTimeSafe,
-} from "../ReservationUnits/ReservationUnitEditor/DateTimeInput";
+} from "app/helpers";
+import ControlledDateInput from "../my-units/components/ControlledDateInput";
 import ControlledTimeInput from "../my-units/components/ControlledTimeInput";
 
 const RichTextInput = dynamic(() => import("app/component/RichTextInput"), {
